Add "Template Name" support to Blade
Add `{{-- Template name: My template --}}` to your Blade template. This
gives the template in WordPress a nice name. It works the same as with
custom WordPress templates:
https://developer.wordpress.org/themes/template-files-section/page-template-files/#creating-custom-page-templates-for-global-use

The Blade template must be inside a directory with the name of the post
type.

The `preg_match` call is copied from the WordPress `get_post_templates()`.